Include the notebook showing the performance of the LC prediction
In this notebook we show the possibility of the QSO light curve prediction by the autoencoder model. The autoencoder is trained by the  80%  of the full length of each light curves, and then we check the accuracy of the prediction by comparing it with the 20% dataset which we didn't use for the training.